Slim query client PR: query() only, FromYdbRow, internal pool.
Remove retry_transaction and public pool API; merge modules to cut diff size.
