+++ ydb/src/grpc_wrapper/raw_query_service/mod.rs
@@ -1,6 +1,4 @@
 pub(crate) mod client;
 pub(crate) mod execute_query;
-pub(crate) mod session;
 pub(crate) mod status;
 pub(crate) mod stream;
-pub(crate) mod transaction_control;
